cloud/aws/cdk/serverless-openai-chatbot/server/lambda/python/login/requirements.txt
PyJWT==2.6.0